fix the confusion between decimal and int: decimal = int float = double also: make string and boolean properties work but there is a problem with the conditions in the second file

+++ simulation-engine/src/engine/entity/impl/EntityInstanceManager.java
@@ -30,7 +30,7 @@ public class EntityInstanceManager {
             switch (currentProperty.getPropertyType()) {
                 case INT:
                     if (currentProperty.getRandomStatus()) {
-                        propertyValue = World.NumberRandomGetter(currentProperty.getFrom(),currentProperty.getTo());
+                        propertyValue = (int) World.NumberRandomGetter(currentProperty.getFrom(),currentProperty.getTo());
                     }
                     newPropertyInstance = new IntProperty((int)propertyValue, currentProperty.getName(), currentProperty.getFrom(), currentProperty.getTo(), currentProperty.getRandomStatus() );
                     break;
